+++ sparschweine/app/server.js
@@ -27,7 +27,7 @@ app.use(express.json());
 app.use('/', require('./routes/sparschweine'));
 app.use('/', require('./routes/kategorien'));
 app.use('/', require('./routes/statistik'));
-app.use('/', require('./routes/import'));
+app.use('/', require('./routes/data'));
 
 app.get('/', sendIndexWithBase);
 app.get('/sparschweine*', sendIndexWithBase);